First pass at splitting up the Router internals.
ScopedRouteCollection has a few too many jobs. It handles both the
building of routes and part of the route parsing/matching. Having the
parsing/matching spread between two classes is a bit gross and indicates
there is a missing class. Separating the building + parsing/route
storage makes a nice break that allows each class to handle one group of
tasks. This division also cleans up Router which is nice.
